services/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-telegram-bot==20.3
asyncpg==0.29.0
qdrant-client==2.7.0
aioredis==2.0.1
PyYAML>=6.0
tiktoken>=0.5.0
# Voice transcription (requires: sudo apt install ffmpeg)
faster-whisper>=1.0.0
# Web search (free, no API key)
duckduckgo-search>=6.0.0
# URL content extraction
html2text>=2020.1.16
